requirements.txt
onnxruntime==1.16.3
numpy==1.24.3
Pillow==10.0.1
flask==2.3.3
opencv-python-headless==4.8.1.78